requirements.txt
absl-py==2.1.0
acres==0.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
etelemetry==0.3.1
executing==2.1.0
fastapi==0.115.8
fastjsonschema==2.20.0
ffmpeg==1.4
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.55.3
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.8.2
google-api-python-client==2.97.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.56.4
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.63.0rc1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.30.2
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
llama-cloud==0.1.13
llama-cloud-services==0.6.3
llama-index==0.12.20
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.20
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.22
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.2
llama_cpp_python==0.3.7
llvmlite==0.44.0
looseversion==1.3.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
more-itertools==10.6.0
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
murmurhash==1.0.12
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
namex==0.0.8
narwhals==1.24.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.24.4
oauthlib==3.2.2
ollama==0.4.7
openai==1.25.0
openai-whisper==20240930
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20250327
pdfplumber==0.11.6
peewee==3.17.8
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.6
plotly==5.24.1
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.0
proto-plus==1.26.0
protobuf==3.20.3
prov==2.0.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
pygame==2.6.1
Pygments==2.19.1
pymongo==10.10.10.10
Pympler==1.1
PyMuPDF==1.25.5
pyparsing==3.2.1
pypdf==5.3.0
pypdfium2==4.30.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-telegram-bot==21.10
pytz==2025.1
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
shap==0.46.0
shellingham==1.5.4
simplejson==3.20.1
simsimd==6.2.1
six==1.17.0
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
streamlit==1.26.0
stringzilla==3.11.3
stripe==11.4.1
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==2.5.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tkcalendar==1.6.1
tokenizers==0.21.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchsummary==1.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.51.2
tweepy==4.14.0
typer==0.15.2
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==4.3.1
ultralytics==8.3.65
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
validators==0.34.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.18.3
yfinance==0.2.51
youtube-search-python==1.6.6
zipp==3.21.0
